requirements.txt
numpy==2.2.2
pandas==2.0.3
scipy==1.10.0
scikit-learn==1.5.1
tqdm==4.66.4
matplotlib==3.7.5
Pillow==10.4.0
pyyaml==6.0.1
requests==2.32.3
orjson==3.10.5
protobuf==4.25.3
torch==2.4.0
torchvision==0.19.0
safetensors==0.4.3
tokenizers==0.19.1
transformers==4.41.2
sentencepiece==0.2.0
joblib==1.4.2
cloudpickle==3.0.0
opencv-python==4.10.0.82
imgaug==0.4.0
scikit-image==0.24.0
open3d==0.18.0
